pyproject.toml
[project]
name = "upstox-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.17.0",
    "python-dotenv>=1.1.1",
    "upstox-python-sdk>=2.18.0",
]
